// src/app/Interfaces/product_interface.ts
export interface IProduct {
  id: string;
  DelFlag: string;
  OrganiztionId: string;
  WebSiteName: string;
  ProductNewURL: string;
  ProductOldURL: string;
  ProductName: string;
  ProductShortDescription: string;
  ProductDetailedDescription: string;
  ProductCategory: string;

  ProductImageURL1: string;
  ProductImageURL2: string;
  ProductImageURL3: string;
  ProductImageURL4: string;
  ProductImageURL5: string;

  ProductAltText1: string;
  ProductAltText2: string;
  ProductAltText3: string;
  ProductAltText4: string;
  ProductAltText5: string;

  ProductSchema: string;

  ProductMetaDataTitle: string;
  ProductMetaDataKeywords: string;
  ProductMetaDataDescription: string;

  ProductRelatedProductId1: string;
  ProductRelatedProductId2: string;
  ProductRelatedProductId3: string;
  ProductRelatedProductId4: string;
  ProductRelatedProductId5: string;

}